refactor(mc): introduce MonteCarloResultView
Compose FanChart, PercentileTable, DepletionHistogram (and the
inline summary cards) into a single result view that any consumer
can render. MonteCarloTab now wires in the controls and delegates
the result rendering. The composite StressTestTab (Task 6.2) will
reuse this same view.
